+++ rulesets.go
@@ -4,7 +4,8 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/url"
-	"time"
+
+	"github.com/ion-channel/ionic/rulesets"
 )
 
 const (
@@ -13,18 +14,7 @@ const (
 	getRuleSetsEndpoint       = "v1/ruleset/getRulesets"
 )
 
-type RuleSet struct {
-	ID          string    `json:"id"`
-	TeamID      string    `json:"team_id"`
-	Name        string    `json:"name"`
-	Description string    `json:"description"`
-	RuleIDs     []string  `json:"rule_ids"`
-	CreatedAt   time.Time `json:"created_at"`
-	UpdatedAt   time.Time `json:"updated_at"`
-	Rules       []Rule    `json:"rules"`
-}
-
-func (ic *IonClient) GetRuleSet(id, teamID string) (*RuleSet, error) {
+func (ic *IonClient) GetRuleSet(id, teamID string) (*rulesets.RuleSet, error) {
 	params := &url.Values{}
 	params.Set("id", id)
 	params.Set("team_id", teamID)
@@ -34,7 +24,7 @@ func (ic *IonClient) GetRuleSet(id, teamID string) (*RuleSet, error) {
 		return nil, fmt.Errorf("failed to get rulesets: %v", err.Error())
 	}
 
-	var rs RuleSet
+	var rs rulesets.RuleSet
 	err = json.Unmarshal(b, &rs)
 	if err != nil {
 		return nil, fmt.Errorf("failed to get rulesets: %v", err.Error())
@@ -43,7 +33,7 @@ func (ic *IonClient) GetRuleSet(id, teamID string) (*RuleSet, error) {
 	return &rs, nil
 }
 
-func (ic *IonClient) GetRuleSets(teamID string) ([]RuleSet, error) {
+func (ic *IonClient) GetRuleSets(teamID string) ([]rulesets.RuleSet, error) {
 	params := &url.Values{}
 	params.Set("team_id", teamID)
 
@@ -52,7 +42,7 @@ func (ic *IonClient) GetRuleSets(teamID string) ([]RuleSet, error) {
 		return nil, fmt.Errorf("failed to get rulesets: %v", err.Error())
 	}
 
-	var rs []RuleSet
+	var rs []rulesets.RuleSet
 	err = json.Unmarshal(b, &rs)
 	if err != nil {
 		return nil, fmt.Errorf("failed to get rulesets: %v", err.Error())
