+++ rulesets/rulesets.go
@@ -1,0 +1,18 @@
+package rulesets
+
+import (
+	"time"
+
+	"github.com/ion-channel/ionic/rules"
+)
+
+type RuleSet struct {
+	ID          string       `json:"id"`
+	TeamID      string       `json:"team_id"`
+	Name        string       `json:"name"`
+	Description string       `json:"description"`
+	RuleIDs     []string     `json:"rule_ids"`
+	CreatedAt   time.Time    `json:"created_at"`
+	UpdatedAt   time.Time    `json:"updated_at"`
+	Rules       []rules.Rule `json:"rules"`
+}
