+++ vulnerabilities.go
@@ -9,7 +9,8 @@ import (
 	"net/http"
 	"net/url"
 	"os"
-	"time"
+
+	"github.com/ion-channel/ionic/vulnerabilities"
 )
 
 const (
@@ -18,36 +19,11 @@ const (
 	getVulnerabilityEndpoint         = "v1/vulnerability/getVulnerability"
 )
 
-// Vulnerability represents a singular vulnerability record in the Ion Channel
-// Platform
-type Vulnerability struct {
-	ID                          int             `json:"id"`
-	ExternalID                  string          `json:"external_id"`
-	Title                       string          `json:"title"`
-	Summary                     string          `json:"summary"`
-	Score                       string          `json:"score"`
-	Vector                      string          `json:"vector"`
-	AccessComplexity            string          `json:"access_complexity"`
-	VulnerabilityAuthentication string          `json:"vulnerability_authentication"`
-	ConfidentialityImpact       string          `json:"confidentiality_impact"`
-	IntegrityImpact             string          `json:"integrity_impact"`
-	AvailabilityImpact          string          `json:"availability_impact"`
-	VulnerabilitySource         string          `json:"vulnerabilty_source"`
-	AssessmentCheck             json.RawMessage `json:"assessment_check"`
-	Scanner                     json.RawMessage `json:"scanner"`
-	Recommendation              string          `json:"recommendation"`
-	ModifiedAt                  time.Time       `json:"modified_at"`
-	PublishedAt                 time.Time       `json:"published_at"`
-	CreatedAt                   time.Time       `json:"created_at"`
-	UpdatedAt                   time.Time       `json:"updated_at"`
-	SourceID                    int             `json:"source_id"`
-}
-
 // GetVulnerabilities returns a slice of Vulnerability for a given product and
 // version string over a specified pagination range.  If version is left blank,
 // it will not be considered in the search query.  An error is returned for
 // client communication and unmarshalling errors.
-func (ic *IonClient) GetVulnerabilities(product, version string, pagination *Pagination) ([]Vulnerability, error) {
+func (ic *IonClient) GetVulnerabilities(product, version string, pagination *Pagination) ([]vulnerabilities.Vulnerability, error) {
 	params := &url.Values{}
 	params.Set("product", product)
 	if version != "" {
@@ -59,7 +35,7 @@ func (ic *IonClient) GetVulnerabilities(product, version string, pagination *Pag
 		return nil, fmt.Errorf("failed to get vulnerabilities: %v", err.Error())
 	}
 
-	var vulns []Vulnerability
+	var vulns []vulnerabilities.Vulnerability
 	err = json.Unmarshal(b, &vulns)
 	if err != nil {
 		return nil, fmt.Errorf("cannot parse vulnerabilities: %v", err.Error())
@@ -72,7 +48,7 @@ func (ic *IonClient) GetVulnerabilities(product, version string, pagination *Pag
 // a slice of vulnerabilities found for the list of dependencies.  An error is
 // returned if the file can't be cannot be read, the API returns an error, or
 // marshalling issues.
-func (ic *IonClient) GetVulnerabilitiesInFile(filePath string) ([]Vulnerability, error) {
+func (ic *IonClient) GetVulnerabilitiesInFile(filePath string) ([]vulnerabilities.Vulnerability, error) {
 	buff := &bytes.Buffer{}
 	bw := multipart.NewWriter(buff)
 
@@ -101,7 +77,7 @@ func (ic *IonClient) GetVulnerabilitiesInFile(filePath string) ([]Vulnerability,
 		return nil, fmt.Errorf("failed to get vulnerabilities: %v", err.Error())
 	}
 
-	var vulns []Vulnerability
+	var vulns []vulnerabilities.Vulnerability
 	err = json.Unmarshal(b, &vulns)
 	if err != nil {
 		return nil, fmt.Errorf("cannot parse vulnerabilities: %v", err.Error())
@@ -112,7 +88,7 @@ func (ic *IonClient) GetVulnerabilitiesInFile(filePath string) ([]Vulnerability,
 
 // GetVulnerability takes an ID string and returns the vulnerability found for
 // that ID.  An error is returned for API errors and marshalling errors.
-func (ic *IonClient) GetVulnerability(id string) (*Vulnerability, error) {
+func (ic *IonClient) GetVulnerability(id string) (*vulnerabilities.Vulnerability, error) {
 	params := &url.Values{}
 	params.Set("external_id", id)
 
@@ -121,7 +97,7 @@ func (ic *IonClient) GetVulnerability(id string) (*Vulnerability, error) {
 		return nil, fmt.Errorf("failed to get vulnerability: %v", err.Error())
 	}
 
-	var vuln Vulnerability
+	var vuln vulnerabilities.Vulnerability
 	err = json.Unmarshal(b, &vuln)
 	if err != nil {
 		return nil, fmt.Errorf("cannot parse vulnerability: %v", err.Error())
